+++ api/tests/unit/application/certification-centers/index_test.js
@@ -1,16 +1,8 @@
-import fs from 'fs';
-import { writeFile, stat, unlink } from 'fs/promises';
-
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
 import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
 import { certificationCenterController } from '../../../../lib/application/certification-centers/certification-center-controller.js';
 
 import * as moduleUnderTest from '../../../../lib/application/certification-centers/index.js';
-import FormData from 'form-data';
-import streamToPromise from 'stream-to-promise';
-
-import * as url from 'url';
-const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
 describe('Unit | Router | certification-center-router', function () {
   describe('GET /api/certification-centers/{certificationCenterId}/divisions', function () {
@@ -270,77 +262,6 @@ describe('Unit | Router | certification-center-router', function () {
       // then
       expect(response.statusCode).to.equal(200);
       sinon.assert.calledOnce(certificationCenterController.findPaginatedSessionSummaries);
-    });
-  });
-
-  describe('POST /api/certification-centers/{certificationCenterId}/sessions/validate-for-mass-import', function () {
-    const testFilePath = `${__dirname}/testFile_temp.csv`;
-
-    let headers;
-    let payload;
-
-    beforeEach(async function () {
-      await writeFile(testFilePath, Buffer.alloc(0));
-      const form = new FormData();
-      const { size: knownLength } = await stat(testFilePath);
-      form.append('file', fs.createReadStream(testFilePath), { knownLength });
-
-      headers = form.getHeaders();
-      payload = await streamToPromise(form);
-    });
-
-    afterEach(async function () {
-      await unlink(testFilePath);
-    });
-
-    it('should exist', async function () {
-      // given
-      sinon.stub(securityPreHandlers, 'checkUserIsMemberOfCertificationCenter').callsFake((_, h) => h.response(true));
-      sinon
-        .stub(securityPreHandlers, 'checkCertificationCenterIsNotScoManagingStudents')
-        .callsFake((_request, h) => h.response(true));
-      sinon.stub(certificationCenterController, 'validateSessionsForMassImport').returns('ok');
-      const certificationCenterId = 123;
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request(
-        'POST',
-        `/api/certification-centers/${certificationCenterId}/sessions/validate-for-mass-import`,
-        payload,
-        null,
-        headers,
-      );
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    context('when user is member of a certification center from a sco organization managing student', function () {
-      it('should forbid access', async function () {
-        // given
-        sinon.stub(securityPreHandlers, 'checkUserIsMemberOfCertificationCenter').callsFake((_, h) => h.response(true));
-        sinon.stub(securityPreHandlers, 'checkCertificationCenterIsNotScoManagingStudents').callsFake((_request, h) => {
-          return Promise.resolve(h.response().code(403).takeover());
-        });
-        sinon.stub(certificationCenterController, 'validateSessionsForMassImport').returns('ok');
-        const certificationCenterId = 123;
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request(
-          'POST',
-          `/api/certification-centers/${certificationCenterId}/sessions/validate-for-mass-import`,
-          payload,
-          null,
-          headers,
-        );
-
-        // then
-        expect(response.statusCode).to.equal(403);
-      });
     });
   });
